move buttons around on search patients form (#2586)
* make changes

* new files

* [pre-commit.ci] auto fixes from pre-commit hooks

* spacing tweaks

---------
