+++ containers/tefca-viewer/src/app/query/test/page.tsx
@@ -6,7 +6,7 @@ import { UseCaseQueryResponse, UseCaseQueryRequest } from "../../query-service";
 // @ts-ignore
 import ResultsView from "../components/ResultsView";
 import MultiplePatientSearchResults from "../components/MultiplePatientSearchResults";
-import SearchForm from "../components/SearchForm";
+import SearchForm from "../components/searchForm/SearchForm";
 import NoPatientsFound from "../components/NoPatientsFound";
 import { Mode, USE_CASES, ValueSetItem } from "../../constants";
